+++ unity/Runtime/Bindings/MjBindings.cs
@@ -6661,6 +6661,9 @@ public static unsafe extern void mj_extractState(mjModel_* m, double* src, uint 
 public static unsafe extern void mj_setState(mjModel_* m, mjData_* d, double* state, uint sig);
 
 [DllImport("mujoco", CallingConvention = CallingConvention.Cdecl)]
+public static unsafe extern void mj_copyState(mjModel_* m, mjData_* src, mjData_* dst, uint sig);
+
+[DllImport("mujoco", CallingConvention = CallingConvention.Cdecl)]
 public static unsafe extern void mj_setKeyframe(mjModel_* m, mjData_* d, int k);
 
 [DllImport("mujoco", CallingConvention = CallingConvention.Cdecl)]